RavenDB-20318 more refactoring
