+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/OngoingTasks/ShardedOngoingTasksHandlerProcessorForGetOngoingTasks.cs
@@ -1,39 +1,20 @@
-﻿using System;
-using System.Net.Http;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using JetBrains.Annotations;
-using Raven.Client.Http;
 using Raven.Server.ServerWide;
-using Raven.Server.Utils;
+using Raven.Server.ServerWide.Context;
 using Raven.Server.Web.Http;
 using Raven.Server.Web.System;
-using Sparrow.Json;
+using Raven.Server.Web.System.Processors.OngoingTasks;
 
 namespace Raven.Server.Documents.Sharding.Handlers.Processors.OngoingTasks
 {
-    internal class ShardedOngoingTasksHandlerProcessorForGetOngoingTasks : ShardedOngoingTasksHandlerProcessorForGetOngoingTasksInfo
+    internal class ShardedOngoingTasksHandlerProcessorForGetOngoingTasks : AbstractOngoingTasksHandlerProcessorForGetOngoingTasks<ShardedDatabaseRequestHandler, TransactionOperationContext>
     {
-        public ShardedOngoingTasksHandlerProcessorForGetOngoingTasks([NotNull] ShardedDatabaseRequestHandler requestHandler) : base(requestHandler)
+        public ShardedOngoingTasksHandlerProcessorForGetOngoingTasks([NotNull] ShardedDatabaseRequestHandler requestHandler) : base(requestHandler, requestHandler.DatabaseContext.OngoingTasks)
         {
         }
 
-        protected override bool SupportsCurrentNode => true;
-
-        protected override RavenCommand<OngoingTasksResult> CreateCommandForNode(string nodeTag)
-        {
-            return new GetOngoingTasksInfoCommand(nodeTag);
-        }
-
-        protected override async ValueTask HandleCurrentNodeAsync()
-        {
-            var result = GetOngoingTasksInternal();
-
-            using (ContextPool.AllocateOperationContext(out JsonOperationContext context))
-            await using (var writer = new AsyncBlittableJsonTextWriterForDebug(context, ServerStore, RequestHandler.ResponseBodyStream()))
-            {
-                context.Write(writer, result.ToJson());
-            }
-        }
+        protected override long SubscriptionsCount => RequestHandler.DatabaseContext.SubscriptionsStorage.GetAllSubscriptionsCount();
 
         protected override Task HandleRemoteNodeAsync(ProxyCommand<OngoingTasksResult> command, OperationCancelToken token)
         {
@@ -41,24 +22,5 @@ namespace Raven.Server.Documents.Sharding.Handlers.Processors.OngoingTasks
                 ? RequestHandler.DatabaseContext.ShardExecutor.ExecuteSingleShardAsync(command, shardNumber, token.Token) 
                 : RequestHandler.DatabaseContext.AllOrchestratorNodesExecutor.ExecuteForNodeAsync(command, command.SelectedNodeTag, token.Token);
         }
-    }
-
-    internal class GetOngoingTasksInfoCommand : RavenCommand<OngoingTasksResult>
-    {
-        public GetOngoingTasksInfoCommand(string nodeTag)
-        {
-            SelectedNodeTag = nodeTag;
-        }
-
-        public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
-        {
-            url = $"{node.Url}/databases/{node.Database}/tasks";
-
-            var request = new HttpRequestMessage { Method = HttpMethod.Get };
-
-            return request;
-        }
-
-        public override bool IsReadRequest => true;
     }
 }
