+++ src/Raven.Server/Web/System/Processors/OngoingTasks/OngoingTasksHandlerProcessorForGetOngoingTasks.cs
@@ -1,39 +1,20 @@
 ﻿using System.Threading.Tasks;
 using JetBrains.Annotations;
-using Raven.Client.Http;
 using Raven.Server.Documents;
-using Raven.Server.Documents.Sharding.Handlers.Processors.OngoingTasks;
 using Raven.Server.ServerWide;
-using Raven.Server.Utils;
+using Raven.Server.ServerWide.Context;
 using Raven.Server.Web.Http;
-using Sparrow.Json;
 
 namespace Raven.Server.Web.System.Processors.OngoingTasks
 {
-    internal class OngoingTasksHandlerProcessorForGetOngoingTasks : OngoingTasksHandlerProcessorForGetOngoingTasksInfo
+    internal class OngoingTasksHandlerProcessorForGetOngoingTasks : AbstractOngoingTasksHandlerProcessorForGetOngoingTasks<DatabaseRequestHandler, DocumentsOperationContext>
     {
-        public OngoingTasksHandlerProcessorForGetOngoingTasks([NotNull] DatabaseRequestHandler requestHandler) : base(requestHandler)
+        public OngoingTasksHandlerProcessorForGetOngoingTasks([NotNull] DatabaseRequestHandler requestHandler) : base(requestHandler, requestHandler.Database.OngoingTasks)
         {
-        }
-
-        protected override async ValueTask HandleCurrentNodeAsync()
-        {
-            var result = GetOngoingTasksInternal();
-
-            using (ContextPool.AllocateOperationContext(out JsonOperationContext context))
-            await using (var writer = new AsyncBlittableJsonTextWriterForDebug(context, ServerStore, RequestHandler.ResponseBodyStream()))
-            {
-                context.Write(writer, result.ToJson());
-            }
         }
 
         protected override Task HandleRemoteNodeAsync(ProxyCommand<OngoingTasksResult> command, OperationCancelToken token) => RequestHandler.ExecuteRemoteAsync(command, token.Token);
-        
-        protected override RavenCommand<OngoingTasksResult> CreateCommandForNode(string nodeTag)
-        {
-            return new GetOngoingTasksInfoCommand(nodeTag);
-        }
-        
-        protected override bool SupportsCurrentNode => true;
+
+        protected override long SubscriptionsCount => RequestHandler.Database.SubscriptionStorage.GetAllSubscriptionsCount();
     }
 }
